% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapflex.R
\name{flex_params}
\alias{flex_params}
\title{List the required parameters for a flexdashboard template}
\usage{
flex_params(template)
}
\arguments{
\item{template}{character, the flexdashboard template. See \code{\link{flex_templates}}.}
}
\value{
a character vector or \code{NULL}
}
\description{
List the required parameters for a flexdashboard template that must be passed to \code{flex}.
}
\details{
If this function returns \code{NULL} for a given template, then no parameters are required to be passed in a named list when calling \code{flex}.
}
\examples{
flex_params("psc1")
}
\seealso{
flex_templates flex
}
