% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapflex.R
\docType{package}
\name{snapflex}
\alias{snapflex}
\alias{snapflex-package}
\title{snapflex: flexdashboards for SNAP data sets.}
\description{
Provides functions and templates for generating local flexdashboards showcasing and summarizing SNAP data sets as well as functions for accessing existing popular SNAP flexdashboards hosted online.
\code{snapflex} is a member package in the SNAPverse collection of R packages.
}
