% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapflex.R
\name{flex_templates}
\alias{flex_templates}
\title{Basic metadata for all flexdashboard templates in snapflex}
\usage{
flex_templates()
}
\value{
a data frame.
}
\description{
This function returns a data frame with basic meta data for all flexdashboard templates in \code{snapflex}.
This includes the following information:
}
\details{
\itemize{
  \item the template ID used for generating a local flexdashboard.
  \item a brief description of the template.
  \item the variables included in the flexdashboard.
  \item whether any variables are displayed as deltas in comparison to a baseline rather than as raw values.
  \item intra- and/or inter-annual time period information if applicable.
  \item whether Shiny is used at runtime to launch a local instance of an interactive flexdashboard rather than a standalone html document.
}
}
\examples{
flex_templates()
}
\seealso{
flex_templates flex_params
}
